requirements.txt
panda
plotly_express
streamlit